updated User.py andded sign up functionality
